Unifica arquivos de utils para cadastro de usuários
